Split Callbacks.cs into separate files per class
