+++ workers/unity/Packages/io.improbable.gdk.core/Subscriptions/Callbacks.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dc5e514e5c4bf4d46ae8610b8ff64567
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
