+++ workers/unity/Packages/io.improbable.gdk.core/Subscriptions/Callbacks/IndexedCallbacks.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7f3c88769a3cd134c8708b1c42ddef12
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
