+++ workers/unity/Packages/io.improbable.gdk.core/Subscriptions/Callbacks/SingleUseIndexCallbacks.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fc7a8fc767133ac4390f7da866da1801
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
